Clean up layer browser example order
