+++ examples/layer-browser/src/examples/index.js
@@ -2,10 +2,12 @@ import coreLayerExamples from './core-layers';
 import experimentalLayerExamples from './experimental-layers';
 // import sampleLayerExamples from './sample-layers';
 import infovisLayerExamples from './infovis-layers';
+import perfLayerExamples from './perf-layers';
 
 export default {
   ...coreLayerExamples,
   ...experimentalLayerExamples,
+  ...infovisLayerExamples,
   // ...sampleLayerExamples,
-  ...infovisLayerExamples
+  ...perfLayerExamples
 };
