+++ modules/experimental-layers/src/index.js
@@ -2,19 +2,24 @@
 // EXPERIMENTAL LAYERS PACKAGE
 //
 
-export {default as MeshLayer} from './mesh-layer/mesh-layer';
-export {default as PathMarkerLayer} from './path-marker-layer/path-marker-layer';
-export {default as PathOutlineLayer} from './path-outline-layer/path-outline-layer';
-
-export {default as Arrow2DGeometry} from './path-marker-layer/arrow-2d-geometry';
-export {default as AdvancedTextLayer} from './advanced-text-layer/advanced-text-layer';
-
+// SHADER MODULES
 export {default as outline} from './shaderlib/outline/outline';
 
-export {default as BezierCurveLayer} from './bezier-curve-layer/bezier-curve-layer';
+// GEOMETRIES
+export {default as Arrow2DGeometry} from './path-marker-layer/arrow-2d-geometry';
 
+// ENHANCED CORE LAYERS
 export {default as GPUGridLayer} from './gpu-grid-layer/gpu-grid-layer';
+export {default as AdvancedTextLayer} from './advanced-text-layer/advanced-text-layer';
 
+//  3D CATALOG LAYERS
+export {default as MeshLayer} from './mesh-layer/mesh-layer';
 export {default as BitmapLayer} from './bitmap-layer/bitmap-layer';
 
+// INFOVIS LAYERS
+export {default as BezierCurveLayer} from './bezier-curve-layer/bezier-curve-layer';
+
+// TRIP LAYERS
 export {default as TripsLayer} from './trips-layer/trips-layer';
+export {default as PathMarkerLayer} from './path-marker-layer/path-marker-layer';
+export {default as PathOutlineLayer} from './path-outline-layer/path-outline-layer';
